fix: component location and include js not work
