+++ src/components/charts/cashEarningChart.tsx
