+++ src/components/charts/earningRevenueChart.tsx
