+++ src/components/charts/pastEarningChart.tsx
