+++ src/pages/past-performance/components/feature/cashEarningChart/buttonGroup.tsx
@@ -12,7 +12,7 @@ function ButtonGroup({ activeButtons, onButtonClick }: ButtonGroupProps) {
     return (
         <div className="flex justify-center">
             <Button
-                className={`rounded-[10px] rounded-r-none bg-white border-2 text-slate-400   text-[12px] hover:bg-gray-200 border-slate-300 ${activeButtons.includes('revenue') ? 'text-black bg-gray-200 poppins-medium' : ''}`}
+                className={`rounded-[10px] rounded-r-none bg-white border-2 text-slate-400   text-[12px] hover:bg-gray-200 border-slate-300 ${activeButtons.indexOf('revenue') !== -1 ? 'text-black bg-gray-200 poppins-medium' : ''}`}
                 onClick={() => onButtonClick('revenue')}
             >
                 <svg width="13" height="12" viewBox="0 0 13 12" className='me-[7px]' fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -21,7 +21,7 @@ function ButtonGroup({ activeButtons, onButtonClick }: ButtonGroupProps) {
                 Revenue
             </Button>
             <Button
-                className={`border-2 border-l-0 rounded-none bg-white text-slate-400 poppins-regular text-[12px] hover:bg-gray-200 border-slate-300 ${activeButtons.includes('earnings') ? 'text-black bg-gray-200 poppins-medium' : ''}`}
+                className={`border-2 border-l-0 rounded-none bg-white text-slate-400 poppins-regular text-[12px] hover:bg-gray-200 border-slate-300 ${activeButtons.indexOf('earnings') !== -1 ? 'text-black bg-gray-200 poppins-medium' : ''}`}
                 onClick={() => onButtonClick('earnings')}
             >
                 <svg width="13" height="12" viewBox="0 0 13 12" className='me-[7px]' fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -31,7 +31,7 @@ function ButtonGroup({ activeButtons, onButtonClick }: ButtonGroupProps) {
             </Button>
 
             <Button
-                className={`border-2 border-l-0 rounded-none bg-white text-slate-400 poppins-regular text-[12px] hover:bg-gray-200 border-slate-300 ${activeButtons.includes('depreciationandamortization') ? 'text-black bg-gray-200 poppins-medium' : ''}`}
+                className={`border-2 border-l-0 rounded-none bg-white text-slate-400 poppins-regular text-[12px] hover:bg-gray-200 border-slate-300 ${activeButtons.indexOf('depreciationandamortization') !== -1 ? 'text-black bg-gray-200 poppins-medium' : ''}`}
                 onClick={() => onButtonClick('depreciationandamortization')}
             >
                 <svg width="13" height="12" viewBox="0 0 13 12" className='me-[7px]' fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -41,7 +41,7 @@ function ButtonGroup({ activeButtons, onButtonClick }: ButtonGroupProps) {
             </Button>
 
             <Button
-                className={`border-2 border-x-0 rounded-none bg-white text-slate-400 poppins-regular text-[12px] hover:bg-gray-200 border-slate-300 ${activeButtons.includes('networkingcapital') ? 'text-black bg-gray-200 poppins-medium' : ''}`}
+                className={`border-2 border-x-0 rounded-none bg-white text-slate-400 poppins-regular text-[12px] hover:bg-gray-200 border-slate-300 ${activeButtons.indexOf('networkingcapital') !== -1 ? 'text-black bg-gray-200 poppins-medium' : ''}`}
                 onClick={() => onButtonClick('networkingcapital')}
             >
                 <svg width="13" height="12" viewBox="0 0 13 12" className='me-[7px]' fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -51,7 +51,7 @@ function ButtonGroup({ activeButtons, onButtonClick }: ButtonGroupProps) {
             </Button>
 
             <Button
-                className={`rounded-[10px] rounded-l-none bg-white border-2 text-slate-400 poppins-regular text-[12px] hover:bg-gray-200 border-slate-300 ${activeButtons.includes('freecashflow') ? 'text-black bg-gray-200 poppins-medium' : ''}`}
+                className={`rounded-[10px] rounded-l-none bg-white border-2 text-slate-400 poppins-regular text-[12px] hover:bg-gray-200 border-slate-300 ${activeButtons.indexOf('freecashflow') !== -1 ? 'text-black bg-gray-200 poppins-medium' : ''}`}
                 onClick={() => onButtonClick('freecashflow')}
             >
                 <svg width="13" height="12" viewBox="0 0 13 12" className='me-[7px]' fill="none" xmlns="http://www.w3.org/2000/svg">
