+++ src/pages/past-performance/components/feature/cashEarningChart/cashEarning.tsx
@@ -5,7 +5,7 @@ import {
     ChartContainer,
     ChartTooltip,
     ChartTooltipContent,
-} from "./chart"
+} from "@/components/charts/cashEarningChart"
 
 const chartData = [
     { type: "Revenue", value: 3, key: "revenue" },
