+++ src/pages/past-performance/components/feature/earningRevenueChart/earningRevenue.tsx
@@ -5,7 +5,7 @@ import {
     ChartContainer,
     ChartTooltip,
     ChartTooltipContent,
-} from "./chart"
+} from "@/components/charts/earningRevenueChart"
 
 
 interface EarningRevenueProps {
