+++ src/pages/past-performance/components/feature/gaugeChart/gauge.tsx
@@ -5,7 +5,7 @@ import {
   ChartContainer,
   ChartTooltip,
   ChartTooltipContent,
-} from "./chart"
+} from "@/components/charts/gaugeChart"
 
 interface GaugeProps {
   title: string;
