+++ src/pages/past-performance/components/feature/pastEarningChart/description.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function description() {
+  return (
+    <div>description</div>
+  )
+}
+
+export default description
