+++ src/pages/past-performance/components/ui/title.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { TitleProps } from '../../type';
+import { TitleProps } from '@/types/index';
 
 function Title({ title, description, svg }: TitleProps) {
     return (
