+++ src/types/index.d.ts
@@ -25,3 +25,32 @@ export interface detailBook  {
 // export interface listCheckedChapters extends detailBook {
 //     bookId: number
 // }
+
+
+// past performance interface
+import { ReactNode } from 'react';
+
+export interface HeaderProps {
+    title: string; 
+}
+
+export interface TitleProps {
+    title: string;
+    description: ReactNode; 
+    svg:ReactNode
+}
+
+export interface ItemProps {
+    id: number,
+    tagId: string;
+    title: string;
+    status: boolean;
+}
+
+export interface HoverState {
+    state: number | null;
+}
+
+export interface CopyRightProps {
+    year: number;
+}
